+++ src/module/advertising/AffiseModuleAdvertisingApi.ts
@@ -1,0 +1,9 @@
+import type { AffiseHasModule } from "../AffiseHasModule";
+
+export interface AffiseModuleAdvertisingApi extends AffiseHasModule {
+    /**
+     * iOS only
+     * Start Advertising module
+     */
+    startModule(): void;
+}